Use UUIDs instead of numeric IDs for StageProp.
This makes it easy for the client to handle undo/redo functionality across saves to the stage. With a typical primary key, adding a stage prop on the client would result in a null ID until it's persisted, which means performing an undo would revert the ID to null, which means a subsequent save will delete the old prop and recreate the new one.